versaai/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & ML (Lightweight versions)
groq==0.4.1
openai==1.3.7
# sentence-transformers==2.8.0
# huggingface_hub==0.20.3
numpy==1.24.3
# scipy==1.11.4  # Commented out - heavy dependency

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring & Logging
loguru==0.7.2
sentry-sdk[fastapi]==1.38.0
psutil==5.9.6

# CORS & Middleware
# Note: FastAPI includes CORS middleware, no additional dependency needed

# Redis for caching (optional)
redis==5.0.1
aioredis==2.0.1
fastapi-cache2==0.2.1

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1